fix: refactor fetches and handle errors slightly better
